Delete tables only when created by own migration

+++ packages/api/cms-api/src/mikro-orm/migrations/Migration20231206123505.ts
@@ -2,8 +2,11 @@ import { Migration } from "@mikro-orm/migrations";
 
 export class Migration20231206123505 extends Migration {
     async up(): Promise<void> {
-        this.addSql('drop table if exists "UserContentScopes";');
-        this.addSql('drop table if exists "UserPermission";');
+        const result = await this.execute(`select count(*) from "Migrations" where name = 'Migration20230831110518';`);
+        if (result[0].count == 1) {
+            this.addSql('drop table if exists "UserContentScopes";');
+            this.addSql('drop table if exists "UserPermission";');
+        }
     }
 
     async down(): Promise<void> {}
